+++ packages/core/index.js
@@ -30,7 +30,8 @@ const create = args => {
     const len = args.length;
     let newStyle = {};
     let id = '';
-    const vars = {};
+    let vars = null;
+    let uses = null;
 
     for (let i = 0; i < len; i++) {
         let style = args[i];
@@ -47,7 +48,11 @@ const create = args => {
         id += '_' + style[KEYS.__id__];
 
         if (style[KEYS.__style__]) {
-            Object.assign(vars, style[KEYS.__style__]);
+            vars = Object.assign(vars || {}, style[KEYS.__style__]);
+        }
+
+        if (style[KEYS.__use__]) {
+            uses = Object.assign(uses || {}, style[KEYS.__use__]);
         }
 
         if (style[KEYS.__store__][id]) {
@@ -68,7 +73,12 @@ const create = args => {
         newStyle[KEYS.__store__] = {
             ['_' + style[KEYS.__id__]]: newStyle,
         };
+    }
+
+    if (vars || use) {
+        newStyle = Object.create(newStyle);
         newStyle[KEYS.__style__] = vars;
+        newStyle[KEYS.__use__] = uses;
     }
 
     return newStyle;
@@ -204,9 +214,11 @@ const appendModifier = (styles, key, value, cn = '') => {
 };
 
 function map(element) {
+    const currStyles = styled[KEYS.__styles__];
     let nextProps = {};
-    let cn = appendElement(styled[KEYS.__styles__], element);
+    let cn = appendElement(currStyles, element);
     let vars = null;
+    let uses = currStyles[KEYS.__use__] || {};
 
     const len = arguments.length;
 
@@ -234,21 +246,38 @@ function map(element) {
 
             const value = currProps[key];
 
-            nextProps[key] = value;
+            cn = appendModifier(currStyles, key, value, cn);
 
-            cn = appendModifier(styled[KEYS.__styles__], key, value, cn);
+            const valueType = typeof value;
+
+            if (
+                valueType === 'string' ||
+                valueType === 'boolean' ||
+                valueType === 'number'
+            ) {
+                const useKey = key + '_' + value;
+                if (key + '_' + true in uses || useKey in uses) {
+                    cn = appendModifier(
+                        currStyles,
+                        USE_PREFIX + key,
+                        value,
+                        cn,
+                    );
+
+                    if (uses[useKey]) {
+                        continue;
+                    }
+                }
+            }
+
+            nextProps[key] = value;
         }
     }
 
     if (useProps) {
         for (let key in useProps) {
             const value = useProps[key];
-            cn = appendModifier(
-                styled[KEYS.__styles__],
-                USE_PREFIX + key,
-                value,
-                cn,
-            );
+            cn = appendModifier(currStyles, USE_PREFIX + key, value, cn);
         }
     }
 
@@ -266,11 +295,8 @@ function map(element) {
     return nextProps;
 }
 
-Object.defineProperty(exports, '__esModule', {
-    value: true,
-});
-
-Object.assign(exports, {
+module.exports = {
+    __esModule: true,
     default: styled,
     use,
     css,
@@ -300,4 +326,4 @@ Object.assign(exports, {
     USE_PREFIX,
     KEYS,
     RESHADOW_ID,
-});
+};