+++ packages/svelte/spec/index.test.js
@@ -16,8 +16,10 @@ describe('svelte preprocess', () => {
             <script>
                 import styled from 'reshadow'
 
+                export let name = 'name'
                 export let disabled = false
                 export let size = 'm'
+                export let variant = 'action'
 
                 styled\`
                     h1 {
@@ -28,9 +30,9 @@ describe('svelte preprocess', () => {
 
             <svelte:body on:click={() => console.log('body click')} />
 
-            <h1>hello world</h1>
+            <h1>hello {name}</h1>
 
-            <button :variant="action" :{size} {disabled}>click me</button>
+            <button :state="pending" :variant={variant} :{size} {disabled}>click me</button>
         `;
 
         expect(code).toMatchSnapshot();